Remove `resultIdentifier` key.
Realized when refactoring SageResearch-iOS to have the navigation and identifiers
match the Kotlin model, that the *only* place where the mapping is
`identifer != resultIdentifier` is the top level Assessment where the schema identifer
may be different. While historically we have *not* had different tasks map to the
same schema, this cannot be said for how our partners are using the framework. As
such, it made more sense to move to explicitly defining the schema identifier but only
for the `Assessment`.