Add flag to allow early call to “readyToSave”
The goal is to allow an Assessment to uploaded results *before* the participant exits the
task. This allows showing a “you’re done” screen to the participant, but if the participant
doesn’t close out the task, their results will still be stored and the end timestamp will be
accurately marked.