Proposed refactor of async actions
In practice, sensors, background measurements, and audio are handled in ways that are
platform specific and can even be specific to a given assessment. In other words, it’s
difficult and obfuscated to attempt to solve for the problems of a specific assessment
in a general way without introducing bugs.

That said, the serialization model needs to be defined in a cross-platform way so the
permission information and asyncronous actions are defined as interfaces within this
framework, while references to async actions have been stripped from navigation.